Final Commit
Finished project, I added two elements to the HTML and pulled them into the script file using their tagname. I created a formula to calculate the average review. I created a table that used information from the arrays to display the rating, condition, the game, ID, etc. I created a few for loops, one of which was used for generating 3 customer reviews.